fixed some oopsies

+++ WebService/WebService/WebService/Models/Transaction.cs
@@ -7,12 +7,12 @@ namespace WebService.Models
 {
     public class Transaction
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FromWallet { get; set; }
         public string ToWallet { get; set; }
         public double Content { get; set; }
 
-        public Transaction(int i, string from, string to, double content) {
+        public Transaction(long i, string from, string to, double content) {
             Id = i;
             FromWallet = from;
             ToWallet = to;
